Rename file KnowRowsHolder -> KnownRowsHolder
